Final Deliverables/auth/requirements.txt
flask
ibm_db
pyjwt
requests
trycourier